refactor: split tests into new file
